Move output_mask into the runner module
