+++ telepresence/runner/output_mask.py
